+++ server/client/src/components/PopUpName.js
@@ -1,8 +1,9 @@
 import React, { Component } from "react";
 import AuthenticationService from "./Authentication";
 import "./css/PopUp.css";
-import { toast } from "react-toastify"
-import 'react-toastify/dist/ReactToastify.css'
+import { toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
+
 export default class PopUpName extends Component {
     constructor(props) {
         super(props);
@@ -15,32 +16,48 @@ export default class PopUpName extends Component {
         this.update = this.update.bind(this);
     }
 
-    update = () => {
+    update = (event) => {
+        // Regex to test for string input
+        const string_re = /^[A-Za-z\s]*$/;
+        let isCorrect = true;
+
         const user = AuthenticationService.getCurrentUser();
         const { firstName, lastName } = this.state;
-        fetch("/profile/updatename", {
-            method: "POST",
-            headers: {
-                Accept: "application/json",
-                "Content-Type": "application/json",
-                authorization: "Bearer " + user.accesstoken,
-            },
-            body: JSON.stringify({
-                firstName: firstName,
-                lastName: lastName,
-                user: user.email,
-            }),
-        })
-            .then((Response) => Response.json())
-            .then((json) => {
-                if (json.token) {
-                    console.log(json.error);
-                    localStorage.clear();
-                    this.props.history.push("/");
-                }
-                toast.success(json.message)
-            });
-        this.props.updateStorage(firstName, lastName);
+
+        if (string_re.test(firstName)) {
+            if (string_re.test(lastName)) {
+                fetch("/profile/updatename", {
+                    method: "POST",
+                    headers: {
+                        Accept: "application/json",
+                        "Content-Type": "application/json",
+                        authorization: "Bearer " + user.accesstoken,
+                    },
+                    body: JSON.stringify({
+                        firstName: firstName,
+                        lastName: lastName,
+                        user: user.email,
+                    }),
+                })
+                    .then((Response) => Response.json())
+                    .then((json) => {
+                        if (json.token) {
+                            console.log(json.error);
+                            localStorage.clear();
+                            this.props.history.push("/");
+                        } else {
+                            toast.success("Name successfully changed!");
+                        }
+                    });
+                this.props.updateStorage(firstName, lastName);
+            } else {
+                event.preventDefault();
+                toast.error("Please enter a valid last name!");
+            }
+        } else {
+            event.preventDefault();
+            toast.error("Please enter a valid first name!");
+        }
     };
 
     handleClick = () => {
@@ -68,6 +85,7 @@ export default class PopUpName extends Component {
                                 className="first_name_input"
                                 type="text"
                                 name="firstName"
+                                required
                                 value={this.state.firstName}
                                 onChange={this.handleChange}
                             />
@@ -79,6 +97,7 @@ export default class PopUpName extends Component {
                                 className="last_name_input"
                                 type="text"
                                 name="lastName"
+                                required
                                 value={this.state.lastName}
                                 onChange={this.handleChange}
                             />
@@ -91,7 +110,7 @@ export default class PopUpName extends Component {
                         />
                     </form>
                 </div>
-            </div >
+            </div>
         );
     }
 }
