+++ server/client/src/components/PopUpPhone.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import { toast } from "react-toastify";
 import AuthenticationService from "./Authentication";
 import "./css/PopUp.css";
 
@@ -13,30 +14,46 @@ export default class PopUpPhone extends Component {
         this.update = this.update.bind(this);
     }
 
-    update = () => {
+    update = (event) => {
+        // Regex to test for number input
+        const number_re = /^[0-9\b]+$/;
+
         const user = AuthenticationService.getCurrentUser();
         const { phone } = this.state;
-        fetch("/profile/updatephone", {
-            method: "POST",
-            headers: {
-                Accept: "application/json",
-                "Content-Type": "application/json",
-                authorization: "Bearer " + user.accesstoken,
-            },
-            body: JSON.stringify({
-                phone: phone,
-                user: user.email,
-            }),
-        })
-            .then((Response) => Response.json())
-            .then((json) => {
-                if (json.error === "TokenExpiredError") {
-                    console.log(json.error);
-                    localStorage.clear();
-                    this.props.history.push("/");
-                }
-            });
-        this.props.updateStorage(phone);
+
+        if (number_re.test(phone)) {
+            if (phone.length == 10) {
+                fetch("/profile/updatephone", {
+                    method: "POST",
+                    headers: {
+                        Accept: "application/json",
+                        "Content-Type": "application/json",
+                        authorization: "Bearer " + user.accesstoken,
+                    },
+                    body: JSON.stringify({
+                        phone: phone,
+                        user: user.email,
+                    }),
+                })
+                    .then((Response) => Response.json())
+                    .then((json) => {
+                        if (json.error === "TokenExpiredError") {
+                            console.log(json.error);
+                            localStorage.clear();
+                            this.props.history.push("/");
+                        } else {
+                            toast.success("Phone number successfully changed!");
+                        }
+                    });
+                this.props.updateStorage(phone);
+            } else {
+                event.preventDefault();
+                toast.error("Please enter a valid phone number!");
+            }
+        } else {
+            event.preventDefault();
+            toast.error("Please enter a valid phone number!");
+        }
     };
 
     handleClick = () => {
@@ -63,6 +80,7 @@ export default class PopUpPhone extends Component {
                             <input
                                 className="phone_input"
                                 type="text"
+                                required
                                 name="phone"
                                 value={this.state.phone}
                                 onChange={this.handleChange}
